fix a pmd warning
